LayoutTests/inspector-obsolete/http-tests/websocket/web-socket-frame-expected.txt

Connected.
Tests that WebSocketFrames are being sent and recieved by Web Inspector.

1- sent: test
1- sent: test
2- sent: exit
1- sent: test
2- sent: exit
3- received: test
1- sent: test
2- sent: exit
3- received: test
4- received: exit
